requirements.txt
Django==3.2.5
django-model-utils==4.1.0
django-widget-tweaks==1.4.8
django-adminlte3==0.1.6
django-environ==0.4.5
whitenoise==5.3.0
django-autoslug==1.9.8
django-debug-toolbar==3.1.1
factory-boy==3.1.0
psycopg2==2.8.6
gunicorn==20.1.0
django-allauth==0.45.0
Pillow==8.3.1